refactor helper functions for launch listener
